requirements.txt
python-dotenv>=1.0.0
playwright>=1.40.0
Pillow>=10.0.0
numpy>=1.26.0
opencv-python-headless>=4.10.0
google-api-python-client>=2.120.0
google-auth>=2.25.0
google-auth-httplib2>=0.2.0
pikepdf>=9.0.0
